+++ src/components/Box.tsx
@@ -1,13 +1,15 @@
+import { CSSProperties } from "react";
 import { useBaseStyleProperties, useElement, useProperty, useRender } from "../hooks";
 import { UseElementResult } from "../hooks/useElement";
 import { BaseStyleProps, Prop } from "../types";
 import { Node } from "./Node";
 
-type BoxProps = BaseStyleProps & {
+export type BoxProps = BaseStyleProps & {
+  className?: string;
+  style?: CSSProperties;
   element?: UseElementResult<HTMLDivElement>;
   color?: Prop<string>;
   children?: React.ReactNode;
-  className?: string;
 };
 
 /**
@@ -17,7 +19,7 @@ type BoxProps = BaseStyleProps & {
  * A rectangle (just a div) with a position, size, angle, and a background color. It can be used to
  * group elements that should be moved as though one.
  */
-export const Box: React.FC<BoxProps> = ({ className, ...props }) => {
+export const Box: React.FC<BoxProps> = ({ className, style, ...props }) => {
   const element = useElement(props.element);
 
   const base = useBaseStyleProperties(props);
@@ -34,7 +36,11 @@ export const Box: React.FC<BoxProps> = ({ className, ...props }) => {
 
   return (
     <Node pos={base.pos}>
-      <div ref={element.ref} className={`absolute ${className || ''}`} style={{ contain: 'content' }}>
+      <div
+        ref={element.ref}
+        className={`absolute ${className || ''}`}
+        style={{ contain: 'content', ...style }}
+      >
         {props.children}
       </div>
     </Node>